test: a user can not delete himself

+++ app/Http/Controllers/Admin/UserController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Spatie\Permission\Models\Role;
+use Illuminate\Support\Facades\Auth;
 
 
 class UserController extends Controller
@@ -111,6 +112,21 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+
+    //Prevenir que un usuario se elimine a si mismo
+
+        if ($user->id === Auth::user()->id) {
+            session()->flash('swal', [
+                'icon' => 'error',
+                'title' => 'Accion denegada',
+                'text' => 'No puedes eliminar tu propia cuenta mientras estás autenticado',
+            ]);
+
+
+            abort(403, 'No puedes eliminar tu propia cuenta mientras estás autenticado');
+            return redirect(route('admin.users.index'));
+        }
+
         //Eliminar roles asociados a un usuario
         $user->roles()->detach();
